style: MyTrip 조회 기능 패키지 변경 - MyProfile -> Trip
