+++ src/main/java/com/triprint/backend/domain/trip/controller/TripController.java
@@ -2,6 +2,10 @@ package com.triprint.backend.domain.trip.controller;
 
 import javax.validation.Valid;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -18,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.triprint.backend.domain.auth.security.CurrentUser;
 import com.triprint.backend.domain.auth.security.UserPrincipal;
 import com.triprint.backend.domain.trip.dto.CreateOrUpdateTripRequest;
+import com.triprint.backend.domain.trip.dto.GetMyTripResponse;
 import com.triprint.backend.domain.trip.dto.GetTripResponse;
 import com.triprint.backend.domain.trip.dto.TripResponse;
 import com.triprint.backend.domain.trip.service.TripService;
@@ -31,6 +36,13 @@ import lombok.RequiredArgsConstructor;
 public class TripController {
 
 	private final TripService tripService;
+
+	@GetMapping("/my")
+	@PreAuthorize("hasRole('ROLE_USER')")
+	public ResponseEntity<Page<GetMyTripResponse>> getMyTripList(@CurrentUser UserPrincipal userPrincipal,
+		@PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable page) {
+		return ResponseEntity.ok(tripService.getMyTripList(userPrincipal.getId(), page));
+	}
 
 	@PostMapping()
 	@PreAuthorize("hasRole('ROLE_USER')")
