+++ src/main/java/com/triprint/backend/domain/user/controller/MyProfileController.java
@@ -2,10 +2,6 @@ package com.triprint.backend.domain.user.controller;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,7 +14,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.triprint.backend.domain.auth.security.CurrentUser;
 import com.triprint.backend.domain.auth.security.UserPrincipal;
-import com.triprint.backend.domain.user.dto.GetMyTripResponse;
 import com.triprint.backend.domain.user.dto.MyProfileImgResponse;
 import com.triprint.backend.domain.user.dto.MyProfileRequest;
 import com.triprint.backend.domain.user.dto.MyProfileResponse;
@@ -32,13 +27,6 @@ import lombok.RequiredArgsConstructor;
 public class MyProfileController {
 
 	private final MyProfileService myProfileService;
-
-	@GetMapping("/trip")
-	@PreAuthorize("hasRole('ROLE_USER')")
-	public ResponseEntity<Page<GetMyTripResponse>> getMyTripList(@CurrentUser UserPrincipal userPrincipal,
-		@PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable page) {
-		return ResponseEntity.ok(myProfileService.getMyTripList(userPrincipal.getId(), page));
-	}
 
 	@GetMapping("/profile")
 	@PreAuthorize("hasRole('ROLE_USER')")
