+++ src/main/java/com/triprint/backend/domain/user/service/MyProfileService.java
@@ -1,15 +1,9 @@
 package com.triprint.backend.domain.user.service;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.triprint.backend.core.exception.ResourceNotFoundException;
-import com.triprint.backend.domain.trip.entity.Trip;
-import com.triprint.backend.domain.trip.repository.TripRepository;
-import com.triprint.backend.domain.user.dto.GetMyTripResponse;
 import com.triprint.backend.domain.user.dto.MyProfileImgResponse;
 import com.triprint.backend.domain.user.dto.MyProfileResponse;
 import com.triprint.backend.domain.user.entity.User;
@@ -23,7 +17,6 @@ public class MyProfileService {
 
 	private final UserRepository userRepository;
 	private final AwsS3Service awsS3Service;
-	private final TripRepository tripRepository;
 
 	public MyProfileResponse getMyProfile(Long memberId) {
 
@@ -77,14 +70,6 @@ public class MyProfileService {
 	public void updateUsername(User user, String username) {
 		user.editUsername(username);
 		userRepository.save(user);
-	}
-
-	public Page<GetMyTripResponse> getMyTripList(Long userId, Pageable page) {
-		User user = userRepository.findById(userId).orElseThrow(() ->
-			new ResourceNotFoundException("일치하는 user 가 없습니다."));
-		Page<Trip> tripList = tripRepository.findByTripUser(user, page);
-
-		return tripList.map((trip) -> new GetMyTripResponse(trip));
 	}
 }
